+++ scripts/add-operators-to-imageset.sh
@@ -132,11 +132,10 @@ fi
 
 redhat_operators=()
 certified_operators=()
-redhat_marketplace=()
 community_operator=()
 
 # Step though all the operator sets and determine which catalog they exist in,
-# with priority order: redhat-operator, certified-operator, redhat-marketplace, community-operator
+# with priority order: redhat-operator, certified-operator, community-operator
 # Operator names are selected from the catalogs in the above catalog order.
 for op_set_name in $(echo $op_sets | tr "," " ")
 do
@@ -150,16 +149,13 @@ do
 		do
 			echo_white -n "$op " >&2  # Keep as echo_white
 
-			# Check if this operator exists in each of the three catalogs
+			# Check if this operator exists in each of the catalogs
 			if grep -q "^$op " .index/redhat-operator-index-v$ocp_ver_major; then
 				[ ! "${op_set_array[$op_set_name]}" ] && redhat_operator+=("#-$op_set_name-operators") && op_set_array[$op_set_name]=1 # A bit of a hack!
 				redhat_operator+=("$op")
 			elif grep -q "^$op " .index/certified-operator-index-v$ocp_ver_major; then
 				[ ! "${op_set_array[$op_set_name]}" ] && certified_operator+=("#-$op_set_name-operators") && op_set_array[$op_set_name]=1
 				certified_operator+=("$op")
-			elif grep -q "^$op " .index/redhat-marketplace-index-v$ocp_ver_major; then
-				[ ! "${op_set_array[$op_set_name]}" ] && redhat_marketplace+=("#-$op_set_name-operators") && op_set_array[$op_set_name]=1
-				redhat_marketplace+=("$op")
 			elif grep -q "^$op " .index/community-operator-index-v$ocp_ver_major; then
 				[ ! "${op_set_array[$op_set_name]}" ] && community_operator+=("#-$op_set_name-operators") && op_set_array[$op_set_name]=1
 				community_operator+=("$op")
@@ -174,7 +170,7 @@ do
 done
 
 # Step though all the operators and determine which catalog they exist in,
-# with priority order: redhat-operator, certified-operator, redhat-marketplace, community-operator
+# with priority order: redhat-operator, certified-operator, community-operator
 # Operator names are selected from the catalogs in the above catalog order.
 if [ "$ops" ]; then
 	declare -A op_set_array
@@ -194,9 +190,6 @@ if [ "$ops" ]; then
 		elif grep -q "^$op " .index/certified-operator-index-v$ocp_ver_major; then
 			[ ! "${op_set_array[$op_set_name]}" ] && certified_operator+=("#-$op_set_name-operators") && op_set_array[$op_set_name]=1
 			certified_operator+=("$op")
-		elif grep -q "^$op " .index/redhat-marketplace-index-v$ocp_ver_major; then
-			[ ! "${op_set_array[$op_set_name]}" ] && redhat_marketplace+=("#-$op_set_name-operators") && op_set_array[$op_set_name]=1
-			redhat_marketplace+=("$op")
 		elif grep -q "^$op " .index/community-operator-index-v$ocp_ver_major; then
 			[ ! "${op_set_array[$op_set_name]}" ] && community_operator+=("#-$op_set_name-operators") && op_set_array[$op_set_name]=1
 			community_operator+=("$op")
@@ -218,7 +211,7 @@ echo >&2
 # Write YAML to output file
 echo "  operators:" >> "$OUTPUT_FILE"
 
-for catalog in redhat_operator certified_operator redhat_marketplace community_operator
+for catalog in redhat_operator certified_operator community_operator
 do
 	list=$(eval echo '${'$catalog'[@]}')   # This is a bit of a hack
 	catalog_name=$(echo $catalog | sed "s/_/-/g")