+++ test/basic-test-using-bundle.sh
@@ -29,12 +29,14 @@ sudo nmcli con up ens224
 cd $TEST_DIR_CONN
 #set +x; bash -c "$(gitrepo=sjbylo/aba; gitbranch=main; curl -fsSL https://raw.githubusercontent.com/$gitrepo/refs/heads/$gitbranch/install)"; set -x
 set +x; bash -c "$(gitrepo=sjbylo/aba; gitbranch=dev; curl -fsSL https://raw.githubusercontent.com/$gitrepo/refs/heads/$gitbranch/install)" -- dev; set -x
-cd aba
+rm -rf aba-vX
+mv aba aba-vX  # test with any repo dir name
+cd aba-vX
 echo cincinnati-operator > templates/operator-set-abatest   # Create a test "operator set"
 
 # Create install bundle
 aba -y bundle --pull-secret '~/.pull-secret.json' --platform vmw --channel fast --version p \
-	--op-sets abatest --ops yaks vault-secrets-operator flux --base-domain example.com \
+	--op-sets abatest --ops yaks nginx-ingress-operator --base-domain example.com \
 	--machine-network 10.0.0.0/20 --dns 10.0.1.8 10.0.2.8 --ntp 10.0.1.8  ntp.example.com --out $TEST_DIR_DISCO/delete-me \
 	$LIGHT
 
@@ -57,7 +59,7 @@ rm -vf delete-me*tar
 cd aba
 ./install
 [ "$LIGHT" ] && aba   # Show the bundle instructions
-[ "$LIGHT" ] && mv -v $TEST_DIR_CONN/aba/mirror/save/mirror_00000*tar $TEST_DIR_DISCO/aba/mirror/save   # Merge the two repos (to save disk space on this filesystem) 
+[ "$LIGHT" ] && mv -v $TEST_DIR_CONN/aba-vX/mirror/save/mirror_00000*tar $TEST_DIR_DISCO/aba/mirror/save   # Merge the two repos (to save disk space on this filesystem) 
 rm -rf $TEST_DIR_CONN/aba   # Not needed anymore
 aba     # Show the bundle instructions
 aba -d mirror load -H $MY_HOST -r -y