+++ test/test5-airgapped-install-local-reg.sh
@@ -89,8 +89,8 @@ vf=~steve/.vmware.conf
 [ ! "$VER_OVERRIDE" ] && VER_OVERRIDE=p
 export VER_OVERRIDE=p  # Must set to p since we do upgrade test below
 [ ! "$oc_mirror_ver_override" ] && oc_mirror_ver_override=v2
-#test-cmd -m "Configure aba.conf for ocp_version '$VER_OVERRIDE'" aba --noask --platform vmw --channel $TEST_CHANNEL --version $VER_OVERRIDE  	# FIXME: put back as it was
-test-cmd -m "Configure aba.conf for ocp_version '$VER_OVERRIDE'" aba --noask --platform vmw --channel fast          --version 4.20  		# FIXME: put back as it was
+test-cmd -m "Configure aba.conf for ocp_version '$VER_OVERRIDE'" aba --noask --platform vmw --channel $TEST_CHANNEL --version $VER_OVERRIDE  	# FIXME: put back as it was
+#test-cmd -m "Configure aba.conf for ocp_version '$VER_OVERRIDE'" aba --noask --platform vmw --channel fast          --version 4.20  		# FIXME: put back as it was
 #test-cmd -m "Show ocp_version in $PWD/aba.conf" "grep -o '^ocp_version=[^ ]*' aba.conf"
 
 # for upgrade tests - reduce the version so it can be upgraded later (see below)
@@ -101,12 +101,15 @@ echo ocp_version=$ocp_version
 ocp_version_desired=$ocp_version  # Get the version from aba.conf since that will be the "latest & previous" version.
 mylog ocp_version_desired is $ocp_version_desired
 ocp_version_major=$(echo $ocp_version_desired | cut -d\. -f1-2)
+mylog ocp_version_major=$ocp_version_major
 ##ocp_version_point=$(echo $ocp_version_desired | cut -d\. -f3)
 #mylog ocp_version_point is $ocp_version_point
 ## Reduce the version to create 'bundle' (below) with 
 source scripts/include_all.sh  # Used to fetch the previous available version
 trap - ERR
-ocp_version_older=$(fetch_all_versions $ocp_channel | tail -2 | head -1)  # Fetch 2nd fron last version
+mylog Showing latest few versions in channel: $ocp_channel
+fetch_all_versions $ocp_channel $ocp_version_major | tail 
+ocp_version_older=$(fetch_all_versions $ocp_channel $ocp_version_major | tail -2 | head -1)  # Fetch 2nd fron last version
 mylog Setting older OCP version to $ocp_version_older
 # CAN BE MISSING # ocp_version_older_point=$(expr $ocp_version_point - 1 )  # Change to one patch version lower # fails if the version below is missing!
 # CAN BE MISSING # ocp_version_older=$ocp_version_major.$ocp_version_older_point
@@ -129,9 +132,6 @@ test-cmd -m "Show ocp_version in $PWD/aba.conf" "grep -o '^ocp_version=[^ ]*' ab
 
 test-cmd -m "Show setting of ask in $PWD/aba.conf" "grep -o '^ask=[^ ]*' aba.conf"
 
-mylog "Setting oc_mirror_version=$oc_mirror_ver_override in aba.conf"
-sed -i "s/^oc_mirror_version=.*/oc_mirror_version=$oc_mirror_ver_override /g" aba.conf
-
 mylog Set up vmware.conf
 test-cmd cp -v $vf vmware.conf 
 sed -i "s#^VC_FOLDER=.*#VC_FOLDER=/Datacenter/vm/abatesting#g" vmware.conf
@@ -151,7 +151,8 @@ source <(normalize-aba-conf)
 mylog "Checking value of: ocp_version=$ocp_version"
 
 # Be sure this file exists
-test-cmd -r 1 30 -m "Init test: download mirror-registry-amd64.tar.gz" "aba --dir test mirror-registry-amd64.tar.gz"
+#test-cmd -r 1 30 -m "Init test: download mirror-registry-amd64.tar.gz" "aba --dir test mirror-registry-amd64.tar.gz"
+test-cmd -r 1 30 -m "Init test: download mirror-registry-amd64.tar.gz" "make -C test mirror-registry-amd64.tar.gz"
 ##! tar tvf test/mirror-registry-amd64.tar.gz && [ -s ~/mirror-registry-amd64.tar.gz ] && cp -v ~/mirror-registry-amd64.tar.gz test
 
 #################################
@@ -191,7 +192,7 @@ mylog "Using container mirror at $reg_host:$reg_port and using reg_ssh_user=$reg
 ### CREATE BUNDLE & COPY TO BASTION ###
 
 test-cmd -h $reg_ssh_user@$int_bastion_hostname -m  "Create test subdir: '$subdir'" "mkdir -v -p $subdir" 
-test-cmd -r 3 3 -m "Creating bundle for channel $TEST_CHANNEL & version $ocp_version, with various operators and extract to '$reg_ssh_user@$int_bastion_hostname:$subdir'" "aba -f bundle --pull-secret '~/.pull-secret.json' --platform vmw --channel fast --version $ocp_version --op-sets abatest --ops web-terminal yaks vault-secrets-operator flux --base-domain example.com --machine-network 10.0.0.0/20 --dns 10.0.1.8 10.0.2.8 --ntp $ntp_ip  ntp.example.com --out - | ssh $reg_ssh_user@$int_bastion_hostname tar -C $subdir -xvf -"
+test-cmd -r 3 3 -m "Creating bundle for channel $TEST_CHANNEL & version $ocp_version, with various operators and extract to '$reg_ssh_user@$int_bastion_hostname:$subdir'" "aba -f bundle --pull-secret '~/.pull-secret.json' --platform vmw --channel $TEST_CHANNEL --version $ocp_version --op-sets abatest --ops web-terminal yaks nginx-ingress-operator flux --base-domain example.com --machine-network 10.0.0.0/20 --dns 10.0.1.8 10.0.2.8 --ntp $ntp_ip  ntp.example.com --out - | ssh $reg_ssh_user@$int_bastion_hostname tar -C $subdir -xvf -"
 
 # Back up the image set conf file so we can upgrade the cluster later
 test-cmd -m "Back up the image set conf file so we can use it to upgrade the cluster later" cp mirror/save/imageset-config-save.yaml mirror/save/imageset-config-save.yaml.release.images
@@ -249,9 +250,9 @@ test-cmd -h $reg_ssh_user@$int_bastion_hostname -m  "List of Operators" "aba --d
 test-cmd -m "Sleep 2m" "read -t 120 xy||true"
 test-cmd -h $reg_ssh_user@$int_bastion_hostname -m  "List of Operators" "aba --dir $subdir/aba/sno run --cmd 'oc get packagemanifests'"
 
-# Test for operators: web-terminal yaks vault-secrets-operator flux
+# Test for operators: web-terminal yaks nginx-ingress-operator flux
 test-cmd -m "Pause, so the operator will show up" "read -t 60 yn || true"
-for op in web-terminal yaks vault-secrets-operator flux
+for op in web-terminal yaks nginx-ingress-operator flux
 do
 	test-cmd -h $reg_ssh_user@$int_bastion_hostname -m  "Ensure $op Operator exists" "aba --dir $subdir/aba/sno run --cmd 'oc get packagemanifests' | grep -i $op"
 done
@@ -264,10 +265,8 @@ mylog Now adding more images to the mirror registry
 
 mylog Runtest: vote-app
 
-# Here, we need to cater for both v1 and v2 of oc-mirror which behave differently
-# For v2, we create a new isc file every time we mirror
-# For v1, we always append to the isc file (not sure what is the best practice).
-[ "$oc_mirror_version" = "v1" ] && gvk=v1alpha2 || gvk=v2alpha1
+# Using oc-mirror v2 (v1 no longer supported)
+gvk=v2alpha1
 # For oc-miror v2 (v2 needs to have only the images that are needed for this next save/load cycle)
 [ -f mirror/save/imageset-config-save.yaml ] && cp -v mirror/save/imageset-config-save.yaml mirror/save/bk.imageset-config-save.yaml.$(date "+%Y-%m-%d-%H:%M:%S")
 # CHANGE ACCUMULATE # if [ "$oc_mirror_version" = "v2" ]; then
@@ -488,17 +487,17 @@ test-cmd -h $reg_ssh_user@$int_bastion_hostname -m  "List of Operators" "aba --d
 test-cmd -m "Sleep 2m" "read -t 120 xy||true"
 test-cmd -h $reg_ssh_user@$int_bastion_hostname -m  "List of Operators" "aba --dir $subdir/aba/$cluster_type run --cmd 'oc get packagemanifests'"
 
-# Test for operators: web-terminal yaks vault-secrets-operator flux
+# Test for operators: web-terminal yaks nginx-ingress-operator flux
 test-cmd -m "Pause, so the operator will show up" "read -t 60 yn || true"
-for op in servicemeshoperator3 #web-terminal yaks vault-secrets-operator flux
+for op in servicemeshoperator3 #web-terminal yaks nginx-ingress-operator flux
 do
 	test-cmd -h $reg_ssh_user@$int_bastion_hostname -m  "Ensure $op Operator exists" "aba --dir $subdir/aba/sno run --cmd 'oc get packagemanifests' | grep -i $op"
 done
 
-# Test for operators: web-terminal yaks vault-secrets-operator flux
-#for op in web-terminal yaks vault-secrets-operator flux
+# Test for operators: web-terminal yaks nginx-ingress-operator flux
+#for op in web-terminal yaks nginx-ingress-operator flux
 test-cmd -m "Pause, so the operator will show up" "read -t 60 yn || true"
-for op in              yaks vault-secrets-operator flux
+for op in              yaks nginx-ingress-operator flux
 do
 	test-cmd -h $reg_ssh_user@$int_bastion_hostname -m  "Ensure $op Operator exists" "aba --dir $subdir/aba/$cluster_type run --cmd 'oc get packagemanifests' | grep -i $op"
 done